conan-r1/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.45.0
peft>=0.12.0
trl>=0.11.0
accelerate>=0.28.0
datasets>=2.18.0
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.11.0
sacrebleu>=2.3.0
nltk>=3.8.0
rouge-score>=0.1.2
hypothesis>=6.100.0
pyyaml>=6.0
tqdm>=4.66.0
Pillow>=10.0.0
